Smooth edge cost when finding nearest neighbours
In SOMap problem, we have edge weights that
 are "the higher - the better". In order
to find K nearest neighbours, we must somehow
transalte those weights to numbers with
meaning "the less - the better".

The first thought here is to take 1 / x as
a new weight for x. Rather, 1000 / x, because
we don't want too small numbers and loss of precision.

But this conversion gives too close results
for neighbours distance, e.g.:
1: 2,1000 30,1000.01 5,1000.01 1014,1000.02 7,1000.02 274,1000.02

So I decided to move to something with more "space" for
difference, so that neighbour differences are bigger.
Smoother function seem appropriate here, so now
the above example looks like this:
1: 2,144.27 30,152.587 76,153.84 274,153.886 298,154.04 5,154.142
266,154.367

Notice that difference is larger, but the set of nearest neighbours
has changed.

We should try different conversion functions later.