Fix problem with file paths in C++ scripts
When invoking C++ executables, e.g. `compute_matrix`,
there was a problem that current directory is the root
of repository, but inside executable it tries to access
some path relative to its source file.

Now, before calling an executable, `cd` to its directory.
This needs to be done with `&&`, because each command
of makefile is run in a separate subshell.