Decouple t-SNE run from obtaining additional info
Previously, we plugged tag names into t-SNE (it
was convenient for outputting). When the need arised
to compute number of posts for each tag, it seemed
natural that this information is NOT needed inside t-SNE.

Therefore, now all additional information (i.e. everything except
nearest neighbour matrix) is computed separately, and only
makes itself available to `get_tiling.py`. As a proof of concept,
number of posts is computed. This finishes #3.

Also, since we now have post counts, this commit implements
coloring tags with respect to relative post count (part of issue #2).